refactor: remove index type export
