--- types/index.ts
@@ -1,2 +1,0 @@
-export type { JoinWithDash } from './join-with-dash'
-export type { SortingNode } from './sorting-node'
