+++ utils/create-node-index-map.ts
@@ -1,6 +1,6 @@
 import type { TSESTree } from '@typescript-eslint/types'
 
-import type { SortingNode } from '../types'
+import type { SortingNode } from '../types/sorting-node'
 
 export let createNodeIndexMap = <Node extends TSESTree.Node>(
   nodes: SortingNode<Node>[],
