+++ utils/get-custom-groups-compare-options.ts
@@ -1,5 +1,5 @@
+import type { SortingNode } from '../types/sorting-node'
 import type { CompareOptions } from './compare'
-import type { SortingNode } from '../types'
 
 interface Options {
   customGroups: Record<string, string[] | string> | CustomGroup[]
