+++ utils/sort-nodes-by-dependencies.ts
@@ -1,6 +1,6 @@
 import type { TSESTree } from '@typescript-eslint/types'
 
-import type { SortingNode } from '../types'
+import type { SortingNode } from '../types/sorting-node'
 
 export interface SortingNodeWithDependencies<
   Node extends TSESTree.Node = TSESTree.Node,
