+++ utils/sort-nodes.ts
@@ -1,5 +1,5 @@
+import type { SortingNode } from '../types/sorting-node'
 import type { CompareOptions } from './compare'
-import type { SortingNode } from '../types'
 
 import { compare } from './compare'
